feat(discover)!: add 'groups' and deprecate 'group'
Add a 'groups' directive in the static discovery specification that
allows nodes to be members of multiple groups. These groups are
deduplicated and created in SMD during discovery.

Deprecate the 'group' directive in the static discovery specification.
As of this commit, 'groups' will still work, but a deprecation warning
is printed if it is used in a node definition. If both 'group' and
'groups' are specified, the 'group' entry is merged with the 'groups'
list as if it were another entry in the list. The 'group' directive will
be removed in a future release.

Signed-off-by: Devon Bautista <17506592+synackd@users.noreply.github.com>