feat(discover): add new format allowing bmc specification
This new discovery format separates the specification of nodes and BMCs.
Nodes are then matched to BMCs based on name or xname. This allows
multiple nodes to be members of the same BMC. Even though this is a
feature of Cray EX hardware, this feature makes the discovery format
more flexible in general, making it so BMC information need not be
specified multiple times.

Here's an example of the new format in YAML:

bmcs:
- name: bmc1
  xname: x1000c0s0b0
  mac: de:ad:be:ef:00:01
  ip: 172.16.0.101
  fqdn: x1000c0s0b0.openchami.cluster
- name: bmc2
  xname: x1000c0s0b1
  mac: de:ad:be:ef:00:02
  ip: 172.16.0.102
  fqdn: x1000c0s0b1.openchami.cluster

nodes:
- name: de01
  nid: 1
  xname: x1000c0s0b0n0
  bmc: bmc1
  groups:
  - compute
  interfaces:
  - mac_addr: de:ca:fb:ad:00:01
    ip_addrs:
    - name: management
      ip_addr: 172.16.0.1
- name: de02
  nid: 2
  xname: x1000c0s0b0n1
  bmc: bmc1
  groups:
  - compute
  interfaces:
  - mac_addr: de:ca:fb:ad:00:02
    ip_addrs:
    - name: management
      ip_addr: 172.16.0.2
- name: de03
  nid: 3
  xname: x1000c0s0b1n0
  bmc: bmc2
  groups:
  - compute
  interfaces:
  - mac_addr: de:ca:fb:ad:00:03
    ip_addrs:
    - name: management
      ip_addr: 172.16.0.3

Signed-off-by: Devon Bautista <17506592+synackd@users.noreply.github.com>